icode

+++ leetcode/17_Letter_Combinations_of_a_Phone_Number.h
@@ -4,7 +4,7 @@ class L17 : public icode
 public:
     void lc(int index, string digits, string item, vector<string> &result)
     {
-        if (index < digits.length())
+        if (index < (int)digits.length())
         {
             int i = 0;
             int num = digits[index] - '0';
